% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct.p23.R
\name{conduct.p23}
\alias{conduct.p23}
\title{Conduct a phase 2/3 dose optimization trial for survival endpoint}
\usage{
conduct.p23(
  data = NULL,
  DCO1 = 16,
  targetEvents2 = c(300, 380),
  dose_selection_endpoint = "ORR",
  method = "Independent Incremental",
  multiplicity.method = "simes",
  e1 = NULL,
  ORRdiff = 0
)
}
\arguments{
\item{data}{Dataset produced by simu.p23trial() function}

\item{DCO1}{Data cutoff for Stage 1}

\item{targetEvents2}{Planned target number of events for Stage 2. Either targetEvents2 must be provided.}

\item{dose_selection_endpoint}{Endpoint for dose selection at Stage 1. "ORR" or "not ORR"}

\item{method}{"simes", "Dunnett".}

\item{ORRdiff}{The tolerated difference between ORR of highest dose and lower doses, with default value 0.}
}
\value{
If method is "Independent Incremental", return an object with variables
\describe{
\item{s}{Selected dose}
\item{z1}{Stage 1 z values for each dose}
\item{z2}{Incremental z statistic at each anaysis at Stage 2, calculated from dose selection to the analysis}
\item{w}{Weight, calculated as sqrt (total events for selected dose + control at stage 1 / targetEvents at each analysis at Stage 2 for the selected dose + control combining stage 1 and stage 2 patients.)}
\item{z.c}{z value for each analysis at Stage 2 for the selected dose}
}
If method is "Disjoint Subjects", return an object with variables
#' \describe{
\item{s}{Selected dose}
\item{z1}{Stage 1 z values for unselected doses combined with the selected dose at kth analysis at Stage 2. z1(k, unselected doses Stage 1 followed by selected dose at k)}
\item{z2}{Incremental z statistic at each anaysis at Stage 2, calculated from dose selection to the analysis}
\item{w}{Weight, calculated as sqrt (total events for selected dose + control at stage 1 / targetEvents at each analysis at Stage 2 for the selected dose + control combining stage 1 and stage 2 patients.)}
}
If method is "Mixture", return an object with variables
\describe{
\item{s}{Selected dose}
\item{z1.unselected}{Stage 1 z values for unselected doses}
\item{z11}{Stage 1 subjects z value at 1st analysis}
\item{z21}{Stage 2 subjects z value at 1st analysis}
\item{omega1}{Weight for 1st analysis combining z11 and z21 to calcualte z1.tilde}
\item{z.tilde}{Test statistics for group sequential design at Stage 2}
\item{z.c}{z value for each analysis at Stage 2 for the selected dose}
\item{z2}{Incremental z statistic at each anaysis at Stage 2, calculated from 1st analysis to each following analysis}
\item{v}{Weight for 2nd and further analysis combining z1.tilde and z2\link{k} to calculate z.tilde\link{k}}
}
}
\description{
This function calculates the stage 2 z statistics according to preplanned target number of events at each analysis.
The function returns (1) Z statistics at Stage 2 based on prespecified target events for pooled stage 1 (selected dose) and stage 2 patients. (2) The function
also returns the weights for combination p value approach in the next step and the weights are sqrt of information fractions.
}
\examples{
#Example (1): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.
 
p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = NULL, rho = NULL, dose_selection_endpoint = "not ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

select_dose_p23(data = p23trial, DCO1 = 16, dose_selection_endpoint = "not ORR")

#Independent incremental; dose not selected by ORR
conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "not ORR",
targetEvents2 = c(300, 380), method = "Independent Incremental")

#Disjoint Subjects; dose not selected by ORR
conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "not ORR",
targetEvents2 = c(300, 380), method = "Disjoint Subjects")

#Mixture; dose not selected by ORR
conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "not ORR",
targetEvents2 = c(300, 380), method = "Mixture")

#Example (2): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on ORR with data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.


p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = c(0.25, 0.3, 0.2, 0.2), rho = 0.7, dose_selection_endpoint = "ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

sel = select_dose_p23(data = p23trial, DCO1 = 16, dose_selection_endpoint = "ORR")

#Independent Incremental, dose selected by ORR
conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "ORR",
targetEvents2 = c(300, 380), method = "Independent Incremental")

#Disjoint Subjects; dose not selected by ORR
d = conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "ORR",
targetEvents2 = c(300, 380), method = "Disjoint Subjects")

#Mixture; dose not selected by ORR
conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "ORR",
targetEvents2 = c(300, 380), method = "Mixture")

#Rejection boundary
bd.z = gsDesign::gsDesign(k=2,alpha=0.025,timing=targetEvents2/targetEvents2[K],sfu=gsDesign::sfLDOF, test.type=1)$upper$bound

#Combination p value
comb.pvalue.p23(z1=matrix(d$z1[1, ], nrow=1),  z2 = d$z2[,1], bd.z=bd.z[1], w=d$w[,1], selected.dose = sel$s, method="simes")
}
