% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.nEachMonth.R
\name{f.nEachMonth}
\alias{f.nEachMonth}
\title{Recruitment Utility Function to Determine the Number of Patients Enrolled per Month}
\usage{
f.nEachMonth(N = 600, A = 24, w = 2, r = 2, Lambda = NULL)
}
\arguments{
\item{N}{Total number patients in two arms.}

\item{A}{Total accrual period in months}

\item{w}{Weight parameter in cumulative enrollment pattern.
The cumulative enrollment at month t is (t / A)^w, eg, at month 6,
the enrollment is N*(6/24)^2 = N/16 for 24 months planned accrual period.}

\item{r}{Randomization ratio r:1, where r refers to the experimental arm, eg, r=2 in 2:1 ratio}

\item{Lambda}{Cumulative distribution function (CDF) for enrollment on (0, infinity). For example, uniform enrollment of 20 patients / month for 24 months has Lambda = function(t){t/24*as.numeric(t<= 24) + as.numeric(t>24)}.}
}
\value{
An object with elements:
\describe{
\item{n0}{number of subjects per month for control arm}
\item{n1}{number of subjects per month for experimental arm}
}
}
\description{
Determine the Number of Patients Enrolled per Month according to specified recruitment pattern.
}
