% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsd.power.R
\name{gsd.power}
\alias{gsd.power}
\title{Calculate power for group sequential design by simulation}
\usage{
gsd.power(z, bd.z, p = NULL, bd.p = NULL)
}
\arguments{
\item{z}{Standard z test statistics. A matrix of dimensions (n_trials, n_analyses).}

\item{bd.z}{z value boundary}

\item{p}{p values (1-sided),i.e., p = 1-pnorm(z)}

\item{bd.p}{p value boundary}
}
\value{
Cumulative power by analysis
}
\description{
This functions calculates the cumulative power and overall power for a group sequential design by simulations.
}
\examples{
#Example (1): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.

#Using O'Brien Fleming boundary, the rejection boundaries are: 
bd.z = actualBounds(planned.events=c(300, 380), act.events=c(300, 380), sf=gsDesign::sfLDOF, alpha=0.025)$actual.z
#2.268527 2.022098

o = simu.ph23data(nSim=1000, n1 = c(50, 50, 50, 50), n2 = c(200, 200), m = c(9, 11, 13, 8), 
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
DCO1 = 16, Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4, targetEvents2 = c(300, 380))

IA = ph23.comb.p(z1=o$z1,  z2 = o$z2[,1], bd.z=bd.z[1], w=o$w[,1])
FA = ph23.comb.p(z1=o$z1,  z2 = o$z2[,2], bd.z=bd.z[2], w=o$w[,2])

#power calculation
gsd.power(z = cbind(IA$comb.z, FA$comb.z), bd.z=bd.z)

}
