% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank.one.sided.R
\name{logrank.one.sided}
\alias{logrank.one.sided}
\title{Perform one-sided logrank test}
\usage{
logrank.one.sided(time, cnsr, group)
}
\arguments{
\item{time}{Survival time}

\item{cnsr}{Censoring status (0 = event, 1 = censor)}

\item{group}{Group indicator (0 = control, 1 = experimental arm)}
}
\value{
\describe{
\item{z}{Test statistics z value}
\item{chisq}{Test statistics chisq value}
\item{two.sided.p}{two.sided p value}
\item{one.sided.p}{one sided p value}
}
}
\description{
This functions performs the one-sided logrank test. The standard logrank test in survival package only produces two-sided test. This function can facilitate one-sided logrank test.
}
\examples{

n = 100
time = c(rexp(n, rate=log(2)/12), rexp(n, rate=log(2)/12*0.7))
cnsr = c(as.numeric(runif(n) > 0.7), as.numeric(runif(n) > 0.7))
group = c(rep(0, n), rep(1, n))

data = as.data.frame(cbind(time, cnsr, group))

logrank.one.sided(time, cnsr, group)

}
