% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph23.comb.p.R
\name{ph23.comb.p}
\alias{ph23.comb.p}
\title{p value combination approach using closed testing procedure (only up to 3 dose levels)}
\usage{
ph23.comb.p(
  z1,
  z2,
  p1 = NULL,
  p2 = NULL,
  bd.z = 1.96,
  w = 0.2,
  method = "simes"
)
}
\arguments{
\item{z1}{A matrix of z values at stage 1. z1 has dimensions of (ntrials, ndoses). ndoses must be 2 or 3.}

\item{z2}{stage 2 z test statistic. a vector of length = ntrials. length(z2) must be equal to nrow(z1)}

\item{p1}{A matrix of p values at stage 1. p1 = 1-pnorm(z1). either z1 or p1 must be provided.}

\item{p2}{stage 2 p value. p2 = 1-pnorm(z2). Either z2 or p2 must be provided.}

\item{bd.z}{rejection boundary in z scale for the combination test}

\item{w}{weight for stage 1 z statistic. If w is a single number, all trials use the same weight in p value combination.}

\item{method}{"simes", "Dunnett". Currently only "simes" method is implemented.}
}
\value{
Returned values include:
\describe{
\item{selected.dose}{Selected dose}
\item{unadj.ps}{unadjusted p value for selected dose s}
\item{adj.ps}{adjusted p value for selected dose s in Stage 1 after closed testing procedure using Simes method}
\item{comb.p}{combination p value combining stage 1 and stage 2}
\item{unadj.zs}{unadjusted z value for selected dose s}
\item{adj.zs}{adjusted z value for selected dose s in Stage 1 after closed testing procedure using Simes method}
\item{comb.z}{Combination z value combining stage 1 and stage 2}
\item{RejPct}{Percentage of rejection, applicable for simulations when z1 has a large number of rows (trials)}
\item{Selection}{Probability of selection of each dose, applicable for simulations when z1 has a large number of rows (trials)}
}
}
\description{
This functions implements p-value combination approach using closed testing procedure for phase 2/3 inferential design with multiple dose levels at Stage 1.
Multiple testing adjustment method is Simes method for each family wise hypothesis H_J at stage 1. Then closed testing procedure is applied for weighted z combining stage 1 and stage 2.
}
\examples{

#Example (1): Stage 1 has 3 z statistics: z1 = c(0.784, 1.118, 1.941)
Select 3rd dose to proceed to Stage 2. At Stage 2, the incremental z test statistic is z2 = 2.1 and the rejection boundary is 1.96.


ph23.comb.p(z1=matrix(c(0.784, 1.118, 1.941), nrow=1),  z2 = 2.1, bd.z=1.96, w=0.2)

ph23.comb.p(z1=matrix(c(0.784, 1.118), nrow=1),  z2 = 2.1, bd.z=1.96, w=0.2)

ph23.comb.p(z1=matrix(c(0.784), nrow=1),  z2 = 2.1, bd.z=1.96, w=0.2)

#Example (2): check type I error by simulations
nSim = 100000
z1 = matrix(NA, nrow=nSim, ncol=3)
for (j in 1:3){z1[ ,j] = rnorm(nSim)}
z2 = rnorm(nSim)

o=ph23.comb.p(z1=z1,  z2 = z2, bd.z=1.96, w=0.2)

#Example (3). w is random between 0 and 1
o=ph23.comb.p(z1=z1,  z2 = z2, bd.z=1.96, w=runif(nSim))

}
