% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dose_p23.R
\name{select_dose_p23}
\alias{select_dose_p23}
\title{Select dose at Stage 1 in a phase 2/3 dose optimization trial for survival endpoint}
\usage{
select_dose_p23(
  data = NULL,
  DCO1 = 16,
  dose_selection_endpoint = "ORR",
  ORRdiff = 0
)
}
\arguments{
\item{data}{Dataset produced by simu.p23trial() function.}

\item{DCO1}{Data cutoff date for stage 1 dose selection}

\item{dose_selection_endpoint}{Dose selection end point: "ORR" or "not ORR"}

\item{ORRdiff}{The tolerated difference between ORR of highest dose and lower doses, with default value 0.}
}
\value{
\describe{
\item{z1}{Stage 1 z values for each dose}
\item{e1}{Stage 1 number of events for each dose}
\item{s}{Selected dose}
}
}
\description{
This function calculates the z statistics (z1) of all arms in stage 1, and select the dose based on maximum z in stage 1.
}
\examples{

#Example (1): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.


p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = NULL, rho = NULL, dose_selection_endpoint = "not ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

select_dose_p23(data = p23trial, DCO1 = 16, dose_selection_endpoint = "not ORR")

#Example (2): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on ORR with data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.


p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = c(0.25, 0.3, 0.4, 0.2), rho = 0.7, dose_selection_endpoint = "ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)


select_dose_p23(data = p23trial, DCO1 = 16, dose_selection_endpoint = "ORR")

}
