% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.single.arm.hz.R
\name{simu.single.arm.hz}
\alias{simu.single.arm.hz}
\title{Simulate a single arm survival data with non-uniform enrollment distribution and drop-off}
\usage{
simu.single.arm.hz(
  n = 100,
  m = 10,
  orr = 0.2,
  rho = 0.7,
  Lambda = function(t) {
     (t/12) * as.numeric(t <= 12) + as.numeric(t > 12)
 },
  A = 12,
  drop = 0,
  DCO = 16,
  targetEvents = NULL
)
}
\arguments{
\item{m}{Median survival time for each arm (dose 1, dose 2, ..., control). length(m) must be equal to length(n1)}

\item{targetEvents}{Planned target number of events for Stage 2. Either targetEvents or DCO must be provided.}

\item{n1}{Sample size of (dose arm, control arm) at Stage 1. length(n1) must be 2.}

\item{n2}{Sample size of the (selected dose arm, control arm). length(n2) must be 2.}

\item{Lambda2}{Enrollment distribution function (CDF) for stage 2.}

\item{Lambda1}{Enrollment distribution function (CDF) for stage 1.}

\item{DCO1}{Data cutoff date for Stage 1}
}
\value{
Datasets including data1: Stage 1 data with multiple dose arms; data2: Stage 2 data of the selected dose arm and control arm with data2[\link{j}] for jth analysis. Each dataset includes variables
\describe{
\item{treatment}{treatment group with values of "control", "dose 1", "dose 2", ...}
\item{enterTime}{Time of randomization in calendar time}
\item{calendarTime}{the time when event/censoring occurred in calendar time}
\item{survTime}{Survival time for analysis, = calendarTime - enterTime}
\item{cnsr}{censor status (0=event; 1=censor) before administrative censoring due to data cut}
\item{calendarCutOff}{Data CutOff Time (DCO);}
\item{survTimeCut}{Survival time after cut}
\item{cnsrCut}{Censor status after cut}
}
The function also returns a dataframe decision with variables:
\describe{
\item{z}{log rank z value for each dose compared to control using stage 1 data}
\item{p}{log rank p value for each dose compared to control using stage 1 data}
\item{adj.p}{adjusted log rank p value for selected dose arm}
\item{adj.z}{adjusted log rank p value for selected dose arm}
\item{selected}{Selected dose arm, applicable for data2 only}
}
}
\description{
This functions simulates trials that have multiple dose arms at Stage 1 and the best dose is always selected after Stage 1 and perform multiple analyses at Stage 2
The function returns multiple datasets including at end of Stage 1 (all dose arms) and at each of analysis at Stage 2
}
\examples{
 
dat=simu.single.arm.hz(n = 100, m = 10, orr = 0.2, Lambda = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A = 12, drop=0, DCO = 16, targetEvents = NULL)

dat=simu.single.arm.hz(n = 100, m = 10, orr = 0.2, Lambda = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A = 12, drop=0, DCO = c(16, 28), targetEvents = NULL)

dat=simu.single.arm.hz(n = 100, m = 10, orr = 0.2, Lambda = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A = 12, drop=0, DCO = NULL, targetEvents = 40)

dat=simu.single.arm.hz(n = 100, m = 10, orr = 0.2, Lambda = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A = 12, drop=0, DCO = NULL, targetEvents = c(40, 60))

}
