% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssr.p23.R
\name{ssr.p23}
\alias{ssr.p23}
\title{Conduct Sample Size Re-estimation for a phase 2/3 dose optimization trial for survival endpoint}
\usage{
ssr.p23(
  data = NULL,
  targetEvents2 = c(300, 380),
  selected.dose = 2,
  ssr_HR_threshold = 0.8,
  events_increase = 30
)
}
\arguments{
\item{data}{Dataset produced by simu.p23trial function}

\item{targetEvents2}{Planned target number of events for Stage 2. Either targetEvents2 must be provided.}

\item{selected.dose}{Selected dose}

\item{ssr_HR_threshold}{HR threshold for SSR. Sample size increase when observed HR at IA in stage 2 is greater than the threshold.}

\item{events_increase}{Number of events planned to increase if the observed HR > the threshold}
}
\value{
If method is "Independent Incremental", return an object with variables
\describe{
\item{targetEvents.FA}{New target number of events at final analysis}
\item{ssr}{Indicator variable whether the sample size is increased.}
}
}
\description{
This function calculates new target events at FA. The SSR is only planned at the first analysis in Stage 2.
}
\examples{
#Example (1): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.
 
p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = NULL, rho = NULL, dose_selection_endpoint = "not ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

sel = select_dose_p23(data = p23trial, DCO1 = 16, dose_selection_endpoint = "not ORR")

#Independent incremental; dose not selected by ORR
ssr.p23(data=p23trial, ssr_HR_threshold = 0.8, events_increase = 30, 
selected.dose=sel$s, targetEvents2 = c(300, 380))

}
