% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct.p23.R
\name{conduct.p23}
\alias{conduct.p23}
\title{Conduct a phase 2/3 dose optimization trial for survival endpoint}
\usage{
conduct.p23(
  data = NULL,
  DCO1 = 16,
  targetEvents2 = c(300, 380),
  dose_selection_endpoint = "ORR",
  method = "Independent Incremental"
)
}
\arguments{
\item{data}{Dataset produced by simu.p23trial() function}

\item{DCO1}{Data cutoff for Stage 1}

\item{targetEvents2}{Planned target number of events for Stage 2. Either targetEvents2 must be provided.}

\item{dose_selection_endpoint}{Endpoint for dose selection at Stage 1. "ORR" or "not ORR"}

\item{method}{"Independent Incremental", "Disjoint Subjects". Currently, only "Independent Incremental" method is implemented.}
}
\value{
\describe{
\item{z1}{Stage 1 z values for each dose}
\item{z.c}{z value for each analysis at Stage 2 for the selected dose}
\item{z2}{Incremental z statistic at each anaysis at Stage 2, calculated from dose selection to the analysis}
\item{w}{Weight, calculated as sqrt (total events for selected dose + control at stage 1 / targetEvents at each analysis at Stage 2 for the selected dose + control combining stage 1 and stage 2 patients.)}
\item{selected.dose}{Selected dose}
\item{example.data}{A simulated dataset example}
}
}
\description{
This function calculates the stage 2 z statistics according to preplanned target number of events at each analysis.
The function returns (1) Z statistics at Stage 2 based on prespecified target events for pooled stage 1 (selected dose) and stage 2 patients. (2) The function
also returns the weights for combination p value approach in the next step and the weights are sqrt of information fractions.
}
\examples{
#Example (1): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.
 
p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = NULL, rho = NULL, dose_selection_endpoint = "not ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

select.dose.p23 (data=p23trial, DCO1=16, dose_selection_endpoint = "not ORR")

conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "not ORR",
targetEvents = c(300, 380), method = "Independent Incremental")

#Example (2): Stage 1: 4 arms; 3 dose levels; each arm 50 patients.
#Stage 2: additional 200 patients per arm will be enrolled at stage 2
#medians for the 4 arms: 9, 11, 13 and control = 8 months
#Enrollment: 12 months uniform in stage 1; 12 months uniform in stage 2
#Holding period: 4 months between stage 1 and 2
#Dose selection will be based on ORR with data cut at 16 months
#Stage 2 has 2 planned analyses at 300 and 380 events respectively.


p23trial = simu.p23trial(n1 = rep(50, 4), n2 = rep(200, 4), m = c(9,9, 9, 9), 
orr = c(0.25, 0.3, 0.2, 0.2), rho = 0.7, dose_selection_endpoint = "ORR",
Lambda1 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A1 = 12,
Lambda2 = function(t){(t/12)*as.numeric(t<= 12) + as.numeric(t > 12)}, A2 = 12,
enrollment.hold=4)

select.dose.p23 (data=p23trial, DCO1=16, dose_selection_endpoint = "ORR")

conduct.p23(data=p23trial, DCO1=16, dose_selection_endpoint = "ORR",
targetEvents = c(300, 380), method = "Independent Incremental")


}
